Refactor commit module to seperate folders
As I’m about to work on dry-runs, I really wanted to split this up because there was just too much going on in this one file
