+++ crates/but/src/command/legacy/status/tui/mod.rs
@@ -8,7 +8,7 @@ use std::{
 use anyhow::Context as _;
 use bstr::BString;
 use but_api::{
-    commit::{commit_insert_blank, ui::RelativeTo},
+    commit::{insert_blank::commit_insert_blank, json::RelativeTo},
     diff::ComputeLineStats,
 };
 use but_ctx::Context;
@@ -510,7 +510,7 @@ impl App {
         }
 
         let reword_result =
-            but_api::commit::commit_reword_only(ctx, commit_id, BString::from(new_message))
+            but_api::commit::reword::commit_reword_only(ctx, commit_id, BString::from(new_message))
                 .with_context(|| format!("failed to reword {}", commit_id.to_hex_with_len(7)))?;
 
         messages.push(Message::Reload(Some(SelectAfterReload::Commit(
@@ -590,9 +590,12 @@ impl App {
             return Ok(());
         }
 
-        let reword_result =
-            but_api::commit::commit_reword_only(ctx, *commit_id, BString::from(new_message))
-                .with_context(|| format!("failed to reword {}", commit_id.to_hex_with_len(7)))?;
+        let reword_result = but_api::commit::reword::commit_reword_only(
+            ctx,
+            *commit_id,
+            BString::from(new_message),
+        )
+        .with_context(|| format!("failed to reword {}", commit_id.to_hex_with_len(7)))?;
 
         messages.extend([
             Message::EnterNormalMode,
